Update all examples to use result builders, other cleanup
Let's make sure that the examples are using result builders when it's helpful to do so.
