+++ Example/EpoxyExample/ViewControllers/FlowLayoutViewController.swift
@@ -40,24 +40,26 @@ final class FlowLayoutViewController: CollectionViewController {
     }
   }
 
-  private var sections: [SectionModel] {
-    [redSection, orangeSection, blueSection]
+  @SectionModelBuilder private var sections: [SectionModel] {
+    redSection
+    orangeSection
+    blueSection
   }
 
   private var redSection: SectionModel {
     // Example of setting insets, item spacing, and line spacing for the section
     SectionModel(
       dataID: DataID.Section.red,
-      items: [
+      items: {
         ColorView.itemModel(
           dataID: DataID.Red.red1,
           style: .red)
-          .flowLayoutItemSize(.init(width: 100, height: 100)),
+          .flowLayoutItemSize(.init(width: 100, height: 100))
         ColorView.itemModel(
           dataID: DataID.Red.red2,
           style: .red)
-          .flowLayoutItemSize(.init(width: 100, height: 100)),
-      ])
+          .flowLayoutItemSize(.init(width: 100, height: 100))
+      })
       .flowLayoutSectionInset(.init(top: 12, left: 20, bottom: 12, right: 48))
       .flowLayoutMinimumInteritemSpacing(18)
       .flowLayoutMinimumLineSpacing(48)
@@ -67,11 +69,11 @@ final class FlowLayoutViewController: CollectionViewController {
     // Example of setting an item size for the entire section
     SectionModel(
       dataID: DataID.Section.orange,
-      items: [
+      items: {
         ColorView.itemModel(
           dataID: DataID.Orange.orange1,
-          style: .orange),
-      ])
+          style: .orange)
+      })
       .flowLayoutItemSize(.init(width: 300, height: 150))
       .flowLayoutSectionInset(.init(top: 0, left: 0, bottom: 0, right: 0))
   }
@@ -80,12 +82,12 @@ final class FlowLayoutViewController: CollectionViewController {
     // Example of setting an item size for the entire section
     SectionModel(
       dataID: DataID.Section.blue,
-      items: [
+      items: {
         ColorView.itemModel(
           dataID: DataID.Blue.item,
           style: .green)
-          .flowLayoutItemSize(.init(width: 200, height: 50)),
-      ])
+          .flowLayoutItemSize(.init(width: 200, height: 50))
+      })
       .supplementaryItems(ofKind: UICollectionView.elementKindSectionHeader, [
         ColorView.supplementaryItemModel(
           dataID: DataID.Blue.header,
